If you want IDE features, use my ENSIME Bundle
I removed all the fancy experimental features I've been working on, this includes:

 - navigation
 - refactoring

The reasoning behind this is simply that my ENSIME bundle does a better job. These two features added a lot of extra dependencies (ctags, scalariform, ruby classes, css, javascript) and we all know that deleted code is easier to maintain. In the ENSIME bundle the ENSIME backend does all the hard work so it's easier for me to maintain.